requirements.txt
equinox==0.11.2
tiktoken==0.5.1
optax==0.1.7
tqdm==4.66.1
transformers==4.35.2
tiktoken==0.5.1
orbax==0.1.9
rich==13.7.0
gcsfs==2023.12.2.post1
wandb==0.16.1
tensorflow-cpu
zstandard==0.22.0
jsonlines==4.0.0
pysimdjson==5.0.2
hydra-core==1.3.2
tensorflow-datasets==4.9.3
tensorstore==0.1.52  # orbax needs 0.1.52 at least
datasets
flax
scikit-learn
jaxopt==0.8.3
cvxpy==1.5.2